observables: Removed LBObservable.
